+++ eagle-core/src/main/java/in/hocg/eagle/modules/account/entity/Account.java
@@ -32,7 +32,7 @@ public class Account extends AbstractEntity<Account> {
      * ID
      */
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
     /**
      * 昵称;显示使用
      */
@@ -97,7 +97,7 @@ public class Account extends AbstractEntity<Account> {
      * 创建者
      */
     @TableField("creator")
-    private Integer creator;
+    private Long creator;
     /**
      * 更新时间
      */
@@ -107,7 +107,7 @@ public class Account extends AbstractEntity<Account> {
      * 更新者
      */
     @TableField("last_updater")
-    private Integer lastUpdater;
+    private Long lastUpdater;
 
 
     @Override
